Fix autolink email detection
@fixed: Autolink now only wraps real URLs/emails instead of every word.
@new: Added build-test man_page_copy target for man page installation.

Technical: Suppressed unused-variable warnings across extensions and renderer.
Technical: Rebuilt build-test artifacts with updated preprocessing logic.

+++ src/extensions/includes.c
@@ -433,7 +433,7 @@ char *apex_process_includes(const char *text, const char *base_dir, int depth) {
                 /* Extract filepath */
                 int filepath_len = filepath_end - filepath_start;
                 char filepath[1024];
-                if (filepath_len > 0 && filepath_len < sizeof(filepath)) {
+                if (filepath_len > 0 && filepath_len < (int)sizeof(filepath)) {
                     memcpy(filepath, filepath_start, filepath_len);
                     filepath[filepath_len] = '\0';
